refactor(delivery): 충돌 부분 리팩토링
