+++ application/order/src/main/java/animal/application/order/domain/delivery/DeliveryStatus.java
@@ -9,7 +9,9 @@ public enum DeliveryStatus {
     WAITING_AT_HUB("허브 대기중"),
     MOVING_TO_HUB("허브 이동중"),
     ARRIVED_AT_HUB("목적지 허브 도착"),
-    DELIVERING("배송중");
+    DELIVERING("배송중"),
+    DELIVERED("배송 완료");
+
 
     private final String status;
 }
