--- application/order/src/main/java/animal/application/order/infrastructure/DeliveryRepository.java
@@ -1,9 +1,0 @@
-package animal.application.order.infrastructure;
-
-import animal.application.order.domain.delivery.Delivery;
-import java.util.UUID;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface DeliveryRepository extends JpaRepository<Delivery, UUID> {
-
-}
