+++ application/order/src/main/java/animal/application/order/infrastructure/order/DeliveryRepository.java
@@ -1,13 +1,12 @@
 package animal.application.order.infrastructure.order;
 
-import animal.application.order.domain.Delivery;
-import animal.application.order.domain.DeliveryId;
+import animal.application.order.domain.delivery.Delivery;
+import animal.application.order.domain.delivery.DeliveryStatus;
 import java.util.List;
 import java.util.UUID;
 import org.springframework.data.jpa.repository.JpaRepository;
-import response.DeliveryStatus;
 
-public interface DeliveryRepository extends JpaRepository<Delivery, DeliveryId> {
+public interface DeliveryRepository extends JpaRepository<Delivery, UUID> {
 
-    List<Delivery> findByHubIdAndDeliveryStatus(UUID hubId, DeliveryStatus deliveryStatus);
+    List<Delivery> findByEndHubIdAndStatus(UUID hubId, DeliveryStatus deliveryStatus);
 }
